requirements.txt
dnspython==2.*